+++ CHANGELOG.md
@@ -4,6 +4,7 @@ X-Core - Changelog
 ### Version 1.0.1-dev
 
 * X-Core Styles: MForm Checkbox wieder ausgerichtet
+* rexx API: `sortArticles()` hinzugefügt, thx@fietstouring
 
 ### Version 1.0.0 - 15. März 2017