+++ CHANGELOG.md
@@ -6,6 +6,7 @@ X-Core - Changelog
 * SCSS Compiler auf Version v0.6.7 aktualisiert
 * X-Core Styles: Extra MForm Tabs Style entfernt
 * X-Core Styles: MForm Checkbox wieder ausgerichtet
+* X-Core Styles: Manche Textareas waren verschwunden (z.B. yForm > Email-Templates) bei ausgeschaltetem Codemirror
 * rexx API: `rexx::sortArticles()` hinzugefügt, thx@fietstouring
 * rexx API: `rexx::formatBytes()` hinzugefügt, gibt Bytes z.b. von `filesize()` formatiert aus z.B. 20,53 MB.
 * rexx API: `rexx::threeWayComparison()` hinzugefügt, PHP5 Ersatz für Spaceship Operator `<=>` aus PHP 7