+++ CHANGELOG.md
@@ -9,11 +9,11 @@ X-Core - Changelog
 * X-Core Styles: Extra MForm Tabs Style entfernt
 * X-Core Styles: MForm Checkbox wieder ausgerichtet
 * X-Core Styles: Manche Textareas waren verschwunden (z.B. yForm > Email-Templates) bei ausgeschaltetem Codemirror
+* Readme: Hinweis zu Unterordner-Installationen hinzugefügt
 * rexx API: `rexx::sortArticles()` hinzugefügt, thx@fietstouring
 * rexx API: `rexx::formatBytes()` hinzugefügt, gibt Bytes z.b. von `filesize()` formatiert aus z.B. 20,53 MB.
 * rexx API: `rexx::threeWayComparison()` hinzugefügt, PHP5 Ersatz für Spaceship Operator `<=>` aus PHP 7
-* rexx API: `rexx::sanitizeFormValue()`, `rexx::validateFormValue()` und weitere Helfer hinzugefügt für den Custom Frontend Form Code
-* Readme: Hinweis zu Unterordner-Installationen hinzugefügt
+* rexx API: `rexx::sanitizeFormValue()`, `rexx::validateFormValue()` und weitere Helfer für den Custom Frontend Form Code hinzugefügt
 
 ### Version 1.0.0 - 15. März 2017